+++ site/dat/docs/Changelog.md
@@ -14,9 +14,16 @@
  - cleanup config from null values before sending it to the cloud
  - pull cumulative percentiles from cloud provisioning
  - replace 'criterias' with 'criteria', add backward compatibility
+ - fix double blazemeter reporting in cloud
  - make ui type configurable with `screen` option of console reporter
  - fix console reporter crash under Windows when curses is installed
  - send some monitoring data into BlazeMeter reporting service
+ - do not clear cloud test files when using Blazemeter reporting
+
+## 1.5.1 <sup>30 may 2016</sup>
+- fix JMeter 3.0 installation issues
+- make JMeter 3.0 the default installed version
+- fix downloading older JMeter versions from Apache archives
 
 ## 1.5.0 <sup>4 may 2016</sup>
  - add [Tsung](Tsung.md) executor